Split up subscribe integration and unit tests in two separate spec files (so CI can skip integration test)
